explains why we use powers of two [FORTRAN]
COMMENTS:
Besides being able to represent these parameters exactly in binary,
having the array size to be an exact power of two means that we
can simplify the implementation of divide-and-conquer algorithms.
We are not building a general purpose application but an specialized
one, thus we can afford to do things as this.