halts execution upon reaching walltime [FORTRAN]
COMMENTS:
The application must comply with the HPCf's jobs requirements, one
of the most relevant ones is the walltime. Since we want a normal
(or graceful) termination the application needs to check if its
runtime does not exceed the alloted time (or the walltime).

For simplicity we use a system-clock with millisecond resolution
since the walltime is about a week time so there's no clock rollover
issue (that would happen in about 25 days, as mentioned in the
GFORTRAN documentation of `system_clock()`. This is the reason
why the walltime is defined in milliseconds in `config`.