dumps status file [FORTRAN]
COMMENTS:
implements auto-scheduling

the submit-script checks the status file and decides what to do
based on its contents, if the status is `pending` the OBDS app
gets submitted again otherwise the script won't do anything.
In that case the status could be `done` or `unknown`.

The `unknown` status is used to initialize the contents of the
status file just in case something goes wrong we are certain that
the submit-script won't submit the job ... otherwise we would
trigger a scheduling hell!

the code has been tested in the trash node

for simplicity we have revised the `sphere::flogger` so that it
returns right after an IO error is detected