adds timer for stopping the OBDS App [FORTRAN]
COMMENTS:
We need a timer that stops the OBDS App once the walltime in the
HPC cluster is reached.

The implementation accounts for the system-clock portability issues.
For example, if the code is compiled with GNU FORTRAN Compiler the
system clock might have a nanosecond resolution whereas if it is
compiled with the Intel FORTRAN Compiler it might have a micro
second resolution. To work around this we simply define the walltime
in seconds and use the system-clock count-rate to perform the
conversion from seconds to `counts`.